Add files via upload
I wasn't able to create the stack due to the error in the screenshot. I looked at the error code and worked on it to fix it, but was unable to. I looked at the Project4 .yml file and it was identical to iteration producing the error, so I don't know what was wrong.